% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cov.R
\name{block.true.cov}
\alias{block.true.cov}
\title{This function construct a covariance matrix with a block diagonal structure.}
\usage{
block.true.cov(p, block.size = 3)
}
\arguments{
\item{p}{The number of variants.}

\item{block.size}{The block size.}
}
\value{
A covariance matrix with a block diagonal structure.
}
\description{
This function construct a covariance matrix with a block diagonal structure.
}
\examples{
data.true.cov <- block.true.cov(30)
data.true.cov[1:9,1:9]
}
